Fix SDL3 migration issues

+++ app/settings/mappingmanager.cpp
@@ -11,6 +11,7 @@
 #define SER_MAPPING "mapping"
 
 MappingFetcher* MappingManager::s_MappingFetcher;
+bool MappingManager::s_AppliedBuiltinMappings;
 
 MappingManager::MappingManager()
 {
@@ -68,34 +69,38 @@ void MappingManager::save()
 
 void MappingManager::applyMappings()
 {
-    QByteArray mappingData = Path::readDataFile("gamecontrollerdb.txt");
-    if (!mappingData.isEmpty()) {
-        int newMappings = SDL_AddGamepadMappingsFromIO(
-                    SDL_IOFromConstMem(mappingData.constData(), mappingData.size()), 1);
+    if (!s_AppliedBuiltinMappings) {
+        QByteArray mappingData = Path::readDataFile("gamecontrollerdb.txt");
+        if (!mappingData.isEmpty()) {
+            int newMappings = SDL_AddGamepadMappingsFromIO(
+                        SDL_IOFromConstMem(mappingData.constData(), mappingData.size()), true);
 
-        if (newMappings > 0) {
-            SDL_LogInfo(SDL_LOG_CATEGORY_APPLICATION,
-                        "Loaded %d new gamepad mappings",
-                        newMappings);
+            if (newMappings > 0) {
+                SDL_LogInfo(SDL_LOG_CATEGORY_APPLICATION,
+                            "Loaded %d new gamepad mappings",
+                            newMappings);
+            }
+            else {
+                if (newMappings < 0) {
+                    SDL_LogError(SDL_LOG_CATEGORY_APPLICATION,
+                                 "Error loading gamepad mappings: %s",
+                                 SDL_GetError());
+                }
+                else if (newMappings == 0) {
+                    SDL_LogWarn(SDL_LOG_CATEGORY_APPLICATION,
+                                "0 new mappings found in gamecontrollerdb.txt. Is it corrupt?");
+                }
+
+                // Try deleting the cached mapping list just in case it's corrupt
+                Path::deleteCacheFile("gamecontrollerdb.txt");
+            }
         }
         else {
-            if (newMappings < 0) {
-                SDL_LogError(SDL_LOG_CATEGORY_APPLICATION,
-                             "Error loading gamepad mappings: %s",
-                             SDL_GetError());
-            }
-            else if (newMappings == 0) {
-                SDL_LogWarn(SDL_LOG_CATEGORY_APPLICATION,
-                            "0 new mappings found in gamecontrollerdb.txt. Is it corrupt?");
-            }
-
-            // Try deleting the cached mapping list just in case it's corrupt
-            Path::deleteCacheFile("gamecontrollerdb.txt");
+            SDL_LogError(SDL_LOG_CATEGORY_APPLICATION,
+                         "Unable to load gamepad mapping file");
         }
-    }
-    else {
-        SDL_LogError(SDL_LOG_CATEGORY_APPLICATION,
-                     "Unable to load gamepad mapping file");
+
+        s_AppliedBuiltinMappings = true;
     }
 
     QList<SdlGamepadMapping> mappings = m_Mappings.values();
